+++ full/src/main/java/apoc/custom/CypherProceduresHandler.java
@@ -58,7 +58,6 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static apoc.ApocConfig.apocConfig;
-import static apoc.custom.CypherProceduresHandlerNewProcedures.qualifiedName;
 import static java.util.Collections.singletonList;
 import static org.neo4j.internal.helpers.collection.MapUtil.map;
 import static org.neo4j.internal.kernel.api.procs.Neo4jTypes.AnyType;
@@ -135,7 +134,7 @@ public class CypherProceduresHandler extends LifecycleAdapter implements Availab
     }
 
     public Mode mode(String s) {
-        return CypherProceduresHandlerNewProcedures.mode(s);
+        return s == null ? Mode.READ : Mode.valueOf(s.toUpperCase());
     }
 
     public Stream<ProcedureOrFunctionDescriptor> readSignatures() {
@@ -275,7 +274,15 @@ public class CypherProceduresHandler extends LifecycleAdapter implements Availab
     }
 
     private String serializeSignatures(List<FieldSignature> signatures) {
-        return CypherProceduresHandlerNewProcedures.serializeSignatures(signatures);
+        List<Map<String, Object>> mapped = signatures.stream().map(fs -> {
+            final Map<String, Object> map = map(
+                    "name", fs.name(),
+                    "type", fs.neo4jType().toString()
+            );
+            fs.defaultValue().map(defVal -> map.put("default", defVal.value()));
+            return map;
+        }).collect(Collectors.toList());
+        return Util.toJson(mapped);
     }
 
     public static List<FieldSignature> deserializeSignatures(String s) {
@@ -421,6 +428,14 @@ public class CypherProceduresHandler extends LifecycleAdapter implements Availab
         }
     }
 
+    public static QualifiedName qualifiedName(@Name("name") String name) {
+        String[] names = name.split("\\.");
+        List<String> namespace = new ArrayList<>(names.length);
+        namespace.add(PREFIX);
+        namespace.addAll(Arrays.asList(names));
+        return new QualifiedName(namespace.subList(0, namespace.size() - 1), names[names.length - 1]);
+    }
+
     public List<FieldSignature> inputSignatures(@Name(value = "inputs", defaultValue = "null") List<List<String>> inputs) {
         List<FieldSignature> inputSignature = inputs == null ? singletonList(FieldSignature.inputField("params", NTMap, DefaultParameterValue.ntMap(Collections.emptyMap()))) :
                 inputs.stream().map(pair -> {
@@ -437,8 +452,7 @@ public class CypherProceduresHandler extends LifecycleAdapter implements Availab
                 outputs.stream().map(pair -> FieldSignature.outputField(pair.get(0), typeof(pair.get(1)))).collect(Collectors.toList());
     }
 
-    // todo - move to NewProcs Handler - because is static
-    public static Neo4jTypes.AnyType typeof(String typeName) {
+    private static Neo4jTypes.AnyType typeof(String typeName) {
         typeName = typeName.replaceAll("\\?", "");
         typeName = typeName.toUpperCase();
         if (typeName.startsWith("LIST OF ")) return NTList(typeof(typeName.substring(8)));
