--- full/src/test/java/apoc/dv/DataVirtualizationCatalogClusterRoutingTest.java
@@ -1,5 +1,0 @@
-package apoc.dv;
-
-
-public class DataVirtualizationCatalogClusterRoutingTest {
-}
