--- full/src/test/java/apoc/dv/DataVirtualizationCatalogNewProceduresTest.java
@@ -1,4 +1,0 @@
-package apoc.dv;
-
-public class DataVirtualizationCatalogNewProceduresTest {
-}
