--- full/src/test/java/apoc/dv/DataVirtualizationCatalogRestartTest.java
@@ -1,4 +1,0 @@
-package apoc.dv;
-
-public class DataVirtualizationCatalogRestartTest {
-}
